requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alpaca-py==0.43.2
annotated-types==0.7.0
attrs==25.4.0
certifi==2026.2.25
charset-normalizer==3.4.4
contourpy==1.3.3
cuda-bindings==12.9.4
cuda-pathfinder==1.4.0
cycler==0.12.1
filelock==3.24.3
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.2.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
lightning-utilities==0.15.3
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
networkx==3.6.1
numpy==2.4.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
packaging==26.0
pandas==3.0.1
pillow==12.1.1
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytorch-lightning==2.6.1
PyYAML==6.0.3
requests==2.32.5
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
setuptools==82.0.0
six==1.17.0
sseclient-py==1.9.0
sympy==1.14.0
ta==0.11.0
threadpoolctl==3.6.0
torch==2.10.0
torchmetrics==1.8.2
tqdm==4.67.3
triton==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
websockets==16.0
yarl==1.22.0
